Changed code from init to morris.py
